move to sever class, because it affects every endpoint
