+++ includes/rest/class-inbox.php
@@ -31,8 +31,6 @@ class Inbox {
 	 */
 	public static function init() {
 		self::register_routes();
-
-		\add_filter( 'rest_request_parameter_order', array( self::class, 'request_parameter_order' ), 10, 2 );
 	}
 
 	/**
@@ -337,33 +335,5 @@ class Inbox {
 		}
 
 		return $users;
-	}
-
-	/**
-	 * Modify the parameter priority order for a REST API request.
-	 *
-	 * @param string[]        $order   Array of types to check, in order of priority.
-	 * @param WP_REST_Request $request The request object.
-	 *
-	 * @return string[] The modified order of types to check.
-	 */
-	public static function request_parameter_order( $order, $request ) {
-		$route = $request->get_route();
-
-		// Check if it is an activitypub request and exclude webfinger and nodeinfo endpoints.
-		if ( ! \str_starts_with( $route, '/' . ACTIVITYPUB_REST_NAMESPACE ) ) {
-			return $order;
-		}
-
-		$type = $request->get_method();
-
-		if ( WP_REST_Server::CREATABLE !== $type ) {
-			return $order;
-		}
-
-		return array(
-			'POST',
-			'defaults',
-		);
 	}
 }
